requirements.txt
streamlit
python-dotenv
PyPDF2
faiss-cpu 
huggingface_hub
torch==2.2.0
transformers==4.42.4
flash-attn==2.2.0
sentence-transformers==2.7.0 
langchain_community